+++ test/unit/Resources/PaymentTypes/PayPalTest.php
@@ -1,0 +1,47 @@
+<?php
+/**
+ * This class defines unit tests to verify functionality of PayPal payment type.
+ *
+ * Copyright (C) 2019 heidelpay GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @link  https://docs.heidelpay.com/
+ *
+ * @author  Simon Gabriel <development@heidelpay.com>
+ *
+ * @package  heidelpayPHP/test/unit
+ */
+namespace heidelpayPHP\test\unit\Resources\PaymentTypes;
+
+use heidelpayPHP\Resources\PaymentTypes\Paypal;
+use heidelpayPHP\test\BaseUnitTest;
+use PHPUnit\Framework\Exception;
+
+class PayPalTest extends BaseUnitTest
+{
+    /**
+     * Verify the bic can be set and read.
+     *
+     * @test
+     *
+     * @throws Exception
+     */
+    public function bicShouldBeRW()
+    {
+        $paypal = new Paypal();
+        $this->assertNull($paypal->getEmail());
+        $paypal->setEmail('test mail');
+        $this->assertEquals('test mail', $paypal->getEmail());
+    }
+}