+++ CHANGELOG.md
@@ -9,9 +9,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 *   Parameters `paymentReference` and `invoiceId` to `Authorization` and `Payout`.
 *   The SDK now supports the webhook event `payout`.
 *   Example for Flexipay direct.
+*   Email parameter to `Paypal` payment type.
 
 ### Changed
 *   The webhook tests now cover all supported events.
+
+### Removed
+*   SAQ-A test due to lack of a corresponding key.
 
 ## [1.2.0.0][1.2.0.0]